Store FT_MOTION linearadvance as int and convert to the real value during processing. Clearer than 10E-6 value

+++ Marlin/src/module/ft_motion.h
@@ -59,7 +59,7 @@ typedef struct FTConfig {
 
   #if HAS_EXTRUDERS
     bool linearAdvEna = FTM_LINEAR_ADV_DEFAULT_ENA;       // Linear advance enable configuration.
-    float linearAdvK = FTM_LINEAR_ADV_DEFAULT_K;          // Linear advance gain.
+    uint16_t linearAdvK = FTM_LINEAR_ADV_DEFAULT_K;          // Linear advance gain.
   #endif
 } ft_config_t;
